requirements.txt
google-genai
django
python-dotenv
markdown2
pandas
pygments
latex2mathml